refactor: binary and directory names change
