+++ Dockerfile
@@ -1,5 +1,5 @@
-# This Dockerfile is for .gorelease purposesBuild: docker build -t seda-chaind .
+# This Dockerfile is for .gorelease purposesBuild: docker build -t sedad .
 # If you are looking for a Dockerfile to run a node, see dockerfiles/Dockerfile.node
 FROM scratch
-ENTRYPOINT ["/seda-chaind"]
-COPY seda-chaind /
+ENTRYPOINT ["/sedad"]
+COPY sedad /
