+++ cmd/seda-chaind/cmd/init.go
@@ -18,7 +18,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/x/genutil/types"
 	"github.com/cosmos/go-bip39"
 
-	"github.com/sedaprotocol/seda-chain/cmd/seda-chaind/utils"
+	"github.com/sedaprotocol/seda-chain/cmd/sedad/utils"
 )
 
 func readInMnemonic(cmd *cobra.Command) (string, error) {
