+++ dockerfiles/Dockerfile.node
@@ -12,7 +12,7 @@ RUN make install
 
 FROM ubuntu:23.04
 EXPOSE 26656 26657 1317 9090
-CMD ["seda-chaind", "start"]
+CMD ["sedad", "start"]
 STOPSIGNAL SIGTERM
 
 
@@ -21,7 +21,7 @@ RUN apt-get update && apt-get install ca-certificates git jq  -y \
 
 
 
-COPY --from=builder /go/bin/seda-chaind /usr/local/bin/
+COPY --from=builder /go/bin/sedad /usr/local/bin/
 COPY --from=builder /go/pkg/mod/github.com/\!cosm\!wasm/wasmvm\@v*/internal/api/libwasmvm.*.so /usr/lib/
 
 COPY scripts/validator_setup/validator_setup.sh /usr/local/bin/validator_setup.sh
