+++ dockerfiles/Dockerfile.static
@@ -35,7 +35,7 @@ RUN ARCH=$(uname -m) && WASMVM_VERSION=$(go list -m github.com/CosmWasm/wasmvm |
 # Copy the remaining files
 COPY . .
 
-# Build seda-chaind binary
+# Build sedad binary
 RUN --mount=type=cache,target=/root/.cache/go-build \
     --mount=type=cache,target=/root/go/pkg/mod \
     GOWORK=off go build \
@@ -43,14 +43,14 @@ RUN --mount=type=cache,target=/root/.cache/go-build \
         -tags "netgo,ledger,muslc" \
         -ldflags \
             "-X github.com/cosmos/cosmos-sdk/version.Name="seda-chain" \
-            -X github.com/cosmos/cosmos-sdk/version.AppName="seda-chaind" \
+            -X github.com/cosmos/cosmos-sdk/version.AppName="sedad" \
             -X github.com/cosmos/cosmos-sdk/version.Version=${GIT_VERSION} \
             -X github.com/cosmos/cosmos-sdk/version.Commit=${GIT_COMMIT} \
             -X github.com/cosmos/cosmos-sdk/version.BuildTags=netgo,ledger,muslc \
             -w -s -linkmode=external -extldflags '-Wl,-z,muldefs -static'" \
         -trimpath \
-        -o /seda-chain/build/seda-chaind \
-        /seda-chain/cmd/seda-chaind/main.go
+        -o /seda-chain/build/sedad \
+        /seda-chain/cmd/sedad/main.go
 
 # --------------------------------------------------------
 # Runner
@@ -58,11 +58,11 @@ RUN --mount=type=cache,target=/root/.cache/go-build \
 
 FROM ${RUNNER_IMAGE}
 
-COPY --from=builder /seda-chain/build/seda-chaind /bin/seda-chaind
+COPY --from=builder /seda-chain/build/sedad /bin/sedad
 
 ENV HOME /seda-chain
 WORKDIR $HOME
 
 EXPOSE 26656 26657 1317 9090
 
-ENTRYPOINT ["seda-chaind"]
+ENTRYPOINT ["sedad"]
