+++ scripts/node_setup/check_validator.sh
@@ -2,12 +2,12 @@
 # Tell script to exit immediately if any cmd fails
 set -e
 
-BIN="seda-chaind"
+BIN="sedad"
 KEY_NAME="${KEY_NAME:-default_key}"
 
 source common.sh
 
 validator_address=$(auth_seda_chaind_command keys show $KEY_NAME --bech val | grep "address:" | awk '{print $3}')
-output=$(seda-chaind query staking validator $validator_address 2>&1)
+output=$(sedad query staking validator $validator_address 2>&1)
 echo "$output" | grep "error:" || echo "$output"
