+++ scripts/testnet/config_example.sh
@@ -5,9 +5,9 @@ GENESIS_TIME="2024-01-18T22:00:00.000000Z"
 CHAIN_VERSION=v0.0.4
 WASMVM_VERSION=v1.5.2
 
-LOCAL_BIN=$(git rev-parse --show-toplevel)/build/seda-chaind # chain binary executable on your machine
+LOCAL_BIN=$(git rev-parse --show-toplevel)/build/sedad # chain binary executable on your machine
 
-HOME_DIR=$HOME/.seda-chain # chain directory
+HOME_DIR=$HOME/.seda # chain directory
 HOME_CONFIG_DIR=$HOME_DIR/config # chain config directory
 
 NODE_DIR=./$CHAIN_ID-nodes # where node directories will be created
